task2_samsung_advisor/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pandas
requests
beautifulsoup4
sentence-transformers
faiss-cpu
python-dotenv